\name{aqlquery}
\alias{aqlquery}
\title{aqlquery}
\usage{
aqlquery(OpenEHRurl, query)
}
\arguments{
  \item{OpenEHRurl}{OpenEHR endpoint}

  \item{query}{AQL query}
}
\description{
A simple openEHR endpoint interface intended
to make results of AQL queries against a chosen openEHR-endpoint 
available as data frame.
}