% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openehR.R
\docType{package}
\name{openehR}
\alias{openehR}
\title{openehR: A package for retrieving data from an openEHR server using the openEHR REST API.}
\description{
`openehR` is an R package for retrieving data from an openEHR server using the
\href{https://specifications.openehr.org/releases/ITS-REST/Release-1.0.1/}{openEHR REST API}.
}
\section{Available functions}{


The available functions are:
\itemize{
\item{\code{\link{query}}}
}
}

