% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openehR.R
\name{query}
\alias{query}
\title{Execute an AQL query on the specified openEHR server.}
\usage{
query(url, query, config = list(), ...)
}
\arguments{
\item{url}{The url of the openEHR server including the openEHR REST API
endpoint}

\item{query}{The AQL query. Newlines and unnecessary whitespaces are stripped
from the query before posting on an openEHR server.}

\item{config}{Optional config to the `httr` package.  (optional)}

\item{...}{additional configuration settings}
}
\value{
the result from the executed query
}
\description{
Execute an AQL query on the specified openEHR server.
}
\examples{
\dontrun{
query("http://localhost:8080/openehr/v1",
     "select c/name/value from composition c limit 5")
query("http://localhost:8080/openehr/v1",
     "select c/name/value from composition c limit 5",
     httr::add_headers(c("name" = "value")))
query("http://localhost:8080/openehr/v1",
     "select c/name/value from composition c limit 5",
     httr::set_cookies(c(a = 1)))
query("http://localhost:8080/openehr/v1",
     "select c/name/value from composition c limit 5",
     httr::authenticate("username", "password"))
query("http://localhost:8080/openehr/v1",
     "select c/name/value from composition c limit 5",
     httr::add_headers(c("name" = "value")),
     httr::authenticate("username", "password"))
query("http://localhost:8080/openehr/v1",
     "select c/name/value from composition c limit 5",
     httr::verbose())
}
}
