\name{query}
\alias{query}
\title{query}
\usage{
query(url, query, config = list())
}
\arguments{
  \item{url}{OpenEHR url}
  \item{query}{AQL query}
  \item{config}{Config}
}
\description{
A simple openEHR endpoint interface intended
to make results of AQL queries against a chosen openEHR-endpoint
available as data frame.
}
